+++ src/test_suite.rs
@@ -1,9 +1,0 @@
-#[cfg(test)]
-mod tests {
-    use crate::{
-        formal_graph_type::FormalGraphType, parser::pg_actions::Property, pg::PropertyGraph,
-        type_map::TypeMap,
-    };
-
-    use super::*;
-}